requirements.txt
flask
numpy
pillow
tflite-runtime
gunicorn